fix: don't print empty pages when they are obviously not needed
Some openpdf quirk is used because the OnPageHandler might not be
called correctly when empty pages are added and later removed from the
document.
The state of the OnPageHandler would then be completely wrong, leading
to missing data in the final document.
Thus, a setting was enabled to also print empty pages.

In case the OnPageHandler has no components to render, this quirk
is not necessary, or rather, it's introducing an undesired behaviour:
when the BaseGenerator is used directly to simply produce some
PDF tables, then unwanted empty pages are generated (depending on
the table's position on the page).

This change only forces to print empty pages when the OnPageHandler
actually controls some components.